LayoutTests/platform/gtk/http/tests/websocket/tests/hybi/extensions-expected.txt

Test WebSocket.extensions attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Check the value before opening the connection:
PASS ws.extensions is ""
Check if the value is read only:
PASS ws.extensions = 'foo'; ws.extensions is ""
Check the value after the connection is established:
FAIL ws.extensions should be x-webkit-deflate-frame. Was permessage-deflate.
PASS successfullyParsed is true

TEST COMPLETE
